requirements.txt
APScheduler==3.6.3
autoflake==1.4
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
mccabe==0.7.0
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
pylama==8.3.8
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.1
tzdata==2022.1
tzlocal==4.1
